♻ data 폴더 생성 & type 경로 이동
