+++ src/components/Chat/ChatList/ChatList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { ChatItem } from 'components';
-import { IChat } from 'pages/ChatListPage/types';
+import { IChat } from 'data/types';
 import * as S from './styles';
 
 interface Prop {
