+++ src/data/chatData.ts
@@ -1,0 +1,20 @@
+export const chatData = [
+  {
+    id: 1,
+    with_user: {
+      name: 'Joy',
+      profile_image_url:
+        'https://user-images.githubusercontent.com/68415905/166093018-2819a713-a7df-4703-bcd5-29b60507bdbf.jpg'
+    },
+    chats: ['안녕하세요', '구매가능한가요?']
+  },
+  {
+    id: 2,
+    with_user: {
+      name: 'Rachel',
+      profile_image_url:
+        'https://user-images.githubusercontent.com/68415905/166093018-2819a713-a7df-4703-bcd5-29b60507bdbf.jpg'
+    },
+    chats: ['안녕하세요']
+  }
+];
