+++ src/data/index.ts
@@ -1,0 +1,1 @@
+export { chatData } from './chatData';
