+++ src/pages/ChatListPage/ChatListPage.tsx
@@ -1,26 +1,6 @@
 import { ChatList, EmptyChat } from 'components';
+import { chatData } from 'data';
 import * as S from './styles';
-
-export const chatData = [
-  {
-    id: 1,
-    with_user: {
-      name: 'Joy',
-      profile_image_url:
-        'https://user-images.githubusercontent.com/68415905/166093018-2819a713-a7df-4703-bcd5-29b60507bdbf.jpg'
-    },
-    chats: ['안녕하세요', '구매가능한가요?']
-  },
-  {
-    id: 2,
-    with_user: {
-      name: 'Rachel',
-      profile_image_url:
-        'https://user-images.githubusercontent.com/68415905/166093018-2819a713-a7df-4703-bcd5-29b60507bdbf.jpg'
-    },
-    chats: ['안녕하세요']
-  }
-];
 
 export const ChatListPage = () => {
   return (
